wip - debug citgm
