+++ test/https/custom-https-server.test.js
@@ -6,55 +6,62 @@ const Fastify = require('../..')
 const https = require('node:https')
 const sget = require('simple-get').concat
 const helper = require('../helper')
-
 const { buildCertificate } = require('../build-certificate')
-t.before(buildCertificate)
 
-test('Should support a custom https server', async t => {
+async function setup () {
+  await buildCertificate()
+
   const localAddresses = await helper.dnsLookup('localhost', { all: true })
-  const minPlan = localAddresses.length - 1 || 1
 
-  t.plan(minPlan + 3)
+  // citgm flaky @ rhel8-s390x rhel8-ppc64le debian10-x64
+  console.log('*** DEBUG dnsLookup(localhost, { all: true }) ***')
+  console.log(localAddresses)
 
-  const serverFactory = (handler, opts) => {
-    t.ok(opts.serverFactory, 'it is called once for localhost')
+  test('Should support a custom https server', { skip: localAddresses.length < 1 }, async t => {
+    t.plan(4)
 
-    const options = {
-      key: global.context.key,
-      cert: global.context.cert
-    }
+    const fastify = Fastify({
+      serverFactory: (handler, opts) => {
+        t.ok(opts.serverFactory, 'it is called once for localhost')
 
-    const server = https.createServer(options, (req, res) => {
-      req.custom = true
-      handler(req, res)
-    })
+        const options = {
+          key: global.context.key,
+          cert: global.context.cert
+        }
 
-    return server
-  }
+        const server = https.createServer(options, (req, res) => {
+          req.custom = true
+          handler(req, res)
+        })
 
-  const fastify = Fastify({ serverFactory })
-
-  t.teardown(fastify.close.bind(fastify))
-
-  fastify.get('/', (req, reply) => {
-    t.ok(req.raw.custom)
-    reply.send({ hello: 'world' })
-  })
-
-  await fastify.listen({ port: 0 })
-
-  await new Promise((resolve, reject) => {
-    sget({
-      method: 'GET',
-      url: 'https://localhost:' + fastify.server.address().port,
-      rejectUnauthorized: false
-    }, (err, response, body) => {
-      if (err) {
-        return reject(err)
+        return server
       }
-      t.equal(response.statusCode, 200)
-      t.same(JSON.parse(body), { hello: 'world' })
-      resolve()
+    })
+
+    t.teardown(fastify.close.bind(fastify))
+
+    fastify.get('/', (req, reply) => {
+      t.ok(req.raw.custom)
+      reply.send({ hello: 'world' })
+    })
+
+    await fastify.listen({ port: 0 })
+
+    await new Promise((resolve, reject) => {
+      sget({
+        method: 'GET',
+        url: 'https://localhost:' + fastify.server.address().port,
+        rejectUnauthorized: false
+      }, (err, response, body) => {
+        if (err) {
+          return reject(err)
+        }
+        t.equal(response.statusCode, 200)
+        t.same(JSON.parse(body), { hello: 'world' })
+        resolve()
+      })
     })
   })
-})
+}
+
+setup()
