+++ test/upgrade.test.js
@@ -7,8 +7,8 @@ const { once } = require('node:events')
 const helper = require('./helper')
 
 async function setup () {
-  const results = await helper.dnsLookup('localhost', { all: true })
-  if (results.length === 1) {
+  const localAddresses = await helper.dnsLookup('localhost', { all: true })
+  if (localAddresses.length === 1) {
     skip('requires both IPv4 and IPv6')
     return
   }
